+++ docker/mollie_entrypoint.sh
@@ -14,4 +14,13 @@ if [ $MOLLIE_TEST_API_KEY != '' ]; then
   sudo service mysql stop;
 fi
 
+
+if [ $MOLLIE_WEBHOOK_DOMAIN != '' ]; then
+    echo "MOLLIE_SHOP_DOMAIN=$MOLLIE_WEBHOOK_DOMAIN" >> /var/www/html/.env
+else
+    # we set a custom webhook domain for the plugin, otherwise payments with localhost do not work ;)
+    echo "MOLLIE_SHOP_DOMAIN=local.mollie.com" >> /var/www/html/.env
+fi
+
+
 exec "$@"