+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [13. February 2022]
+
+### Added
+
+- Add new environment variable **MOLLIE_WEBHOOK_DOMAIN** to set a custom webhook URL.
+
 ## [12. February 2022]
 
 ### Added